requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dnspython==1.16.0
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
lxml==4.4.0
MarkupSafe==1.1.1
PyJWT==1.7.1
pymongo==3.8.0
PySocks==1.7.0
pytz==2019.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
twilio==6.29.3
urllib3==1.25.3
Werkzeug==0.15.5
